% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{adjust_indices_after_remove}
\alias{adjust_indices_after_remove}
\title{Adjust Indices in a List}
\usage{
adjust_indices_after_remove(indicesList, removedIndices)
}
\arguments{
\item{indicesList}{A list of integer vectors, each representing a set of indices.}

\item{removedIndices}{A vector of integers representing the indices to be removed.}
}
\value{
A list of adjusted indices. Each set of indices is adjusted separately.
}
\description{
This function adjusts a list of indices after certain indices have been removed.
The new indices will point to the same elements as the original indices.
If an index is removed, it will also be removed from the indices list.
}
