% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{AIC.ldt.estim}
\alias{AIC.ldt.estim}
\title{Akaike Information Criterion}
\usage{
\method{AIC}{ldt.estim}(object, ..., k = NA)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}}

\item{...}{Additional arguments.}

\item{k}{Unused parameter.}
}
\value{
The value of AIC for the whole system.
#' @importFrom stats AIC
}
\description{
This function extracts Akaike information criterion from an \code{ldt.estim} object.
}
