% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{BIC.ldt.estim}
\alias{BIC.ldt.estim}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{ldt.estim}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}}

\item{...}{Additional arguments.}
}
\value{
The value of BIC for the whole system.
}
\description{
This function extracts Baysian information criterion from an \code{ldt.estim} object.
}
