% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BindVariables}
\alias{BindVariables}
\title{Binds a List of Variables}
\usage{
BindVariables(varList)
}
\arguments{
\item{varList}{A list of variables ((i.e., \code{ldtv} objects)) with similar frequency class}
}
\value{
A matrix with variables in the columns and frequencies as the row names.
}
\description{
Binds a List of Variables
}
\examples{
v1 = ldt::Variable(c(1,2,3,2,3,4,5),"V1",F_Monthly(2022,12), list())
v2 = ldt::Variable(c(10,20,30,20,30,40,50),"V2",F_Monthly(2022,8), list())
vs = ldt::BindVariables(list(v1,v2))
}
