% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClusterH}
\alias{ClusterH}
\title{Hierarchical Clustering}
\usage{
ClusterH(distances, numVariables, linkage = "single")
}
\arguments{
\item{distances}{(numeric vector) Determines the distances. This must be the lower triangle of a (symmetric) distance matrix (without the diagonal).}

\item{numVariables}{(int) Determines the number of variables. This should hold: '2 * length(\code{distances}) = \code{numVariables}(\code{numVariables} - 1)'.}

\item{linkage}{(string) Determines how Distances are calculated in a left-right node merge. It can be \code{single}, \code{complete}, \code{uAverage}, \code{wAverage}, \code{ward}.}
}
\value{
A list:
\item{merge}{(integer matrix)}
\item{height}{(numeric vector)}
\item{order}{(integer vector)}
}
\description{
Hierarchical Clustering
}
