% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{coef.ldt.estim}
\alias{coef.ldt.estim}
\title{Extract Coefficients Matrix}
\usage{
\method{coef}{ldt.estim}(object, equations = NULL, removeZeroRest = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}}

\item{equations}{A number, a numeric array or a string array specifying the indices or names of the endogenous variables in the equations.
\code{NULL} means all equations.}

\item{removeZeroRest}{If \code{TRUE} and model is restricted, zero restrictions are removed.}

\item{...}{Additional arguments.}
}
\value{
If zero restrictions are not removed, it is a matrix containing the coefficients of the system.
Each column of the matrix belongs to an equation. Explanatory variables are in the rows.
Otherwise, coefficients of different equations are reported in a list.
}
\description{
This function extracts coefficient matrix from an \code{ldt.estim} object.
}
