% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{combine.search}
\alias{combine.search}
\title{Combine a List of \code{ldt.search} Objects}
\usage{
combine.search(list, method)
}
\arguments{
\item{list}{A list of \code{ldt.search} objects}

\item{method}{Method in objects}
}
\value{
the combined \code{ldtsearch} object
}
\description{
Combine a List of \code{ldt.search} Objects
}
