% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combineSearch}
\alias{combineSearch}
\title{Combine More Than One \code{ldtsearch} Objects}
\usage{
combineSearch(list, type1Name = "coefs")
}
\arguments{
\item{list}{a list with \code{ldtsearch} objects}

\item{type1Name}{the name of 'type1' in the object}
}
\value{
the combined \code{ldtsearch} object
}
\description{
Combine More Than One \code{ldtsearch} Objects
}
