% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldt_init.R
\docType{data}
\name{data.berka}
\alias{data.berka}
\title{Berka and Sochorova (1993) Dataset for Loan Default}
\format{
A list with the following items:
\describe{
\item{y}{A vector representing the labels. It contains 1 for default and 0 for non-default. Any observation with default (i.e., 'default' and 'finished with default') is considered to be a positive.}
\item{x}{A matrix with explanatory variables in the columns.}
\item{w}{A vector with the weight of each observation. This is mathematically generated to balance the observations.}
\item{descriptions}{A list that describes the column names.}
}
}
\source{
\insertCite{datasetBerka;textual}{ldt}
}
\usage{
data.berka
}
\description{
This dataset is a part of the Berka and Sochorova (1993) study, which contains information on loan defaults. The data was generated using the code in the '/data-raw/data-berka.R' file.
}
\keyword{datasets}
