% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldt_init.R
\docType{data}
\name{data.pcp}
\alias{data.pcp}
\title{IMF's Primary Commodity Prices}
\format{
A list with the following items:
\describe{
\item{data}{A data frame with monthly variables in the columns.}
\item{descriptions}{A list that describes the columns of \code{data}.}
\item{datatypes}{A character array that describes the type of data in the columns of \code{data}.}
\item{start}{A number that indicates the frequency of the first observation in \code{data}.}
}
}
\source{
\insertCite{datasetPcp;textual}{ldt}
}
\usage{
data.pcp
}
\description{
This is a subset of the IMF's Primary Commodity Prices dataset (non-index data is omitted).
The data was generated using the code in the '/data-raw/data-pcp.R' file.
}
\keyword{datasets}
