% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldt_init.R
\docType{data}
\name{data.wdi}
\alias{data.wdi}
\title{Long-run Growth from World Development Indicator Dataset}
\format{
A list with the following items:
\describe{
\item{y}{A vector representing the long-run output growth after 2006. Each element represents a country.}
\item{x}{A matrix with explanatory variables in the columns. Each row represents a country.}
\item{splitYear}{A number that indicates how \code{y} and \code{x} are calculated.}
\item{names}{A list of pairs that describe the code and the name of variables in \code{y} and \code{x}.}
}
}
\source{
\insertCite{datasetWorldbank;textual}{ldt}
}
\usage{
data.wdi
}
\description{
This dataset is derived from the World Development Indicator (WDI) dataset. It contains information on long-run output growth after 2006 and its potential explanatory variables before that year. The data was generated using the code in the '/data-raw/data-wdi.R' file.
}
\keyword{datasets}
