% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Data_Pcp}
\alias{Data_Pcp}
\title{Use 'PCP' Data (i.e., 'IMF's Primary Commodity Prices') and create \code{Date-Series} Table}
\usage{
Data_Pcp(dirPath, makeReal = FALSE)
}
\arguments{
\item{dirPath}{path to the downloaded data data.
It must also contain a file with the US CPI.}

\item{makeReal}{uses the first column (which must be US-CPI) and
converts nominal variables to real}
}
\value{
a list with data, descriptions, etc.
}
\description{
Use 'PCP' Data (i.e., 'IMF's Primary Commodity Prices') and create \code{Date-Series} Table
}
