% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{endogenous}
\alias{endogenous}
\title{Extract Endogenous Variable(s) Data}
\usage{
endogenous(object, equations = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}.}

\item{equations}{A number, a numeric array or a string array specifying the indices or names of the endogenous variables in the equations. \code{NULL} means all equations.}

\item{...}{Additional arguments.}
}
\value{
A data frame containing the endogenous data.
}
\description{
This function extracts data of a endogenous variable(s) from an estimated model.
}
