% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{estim.binary.model.string}
\alias{estim.binary.model.string}
\title{Get Model Name}
\usage{
estim.binary.model.string(object)
}
\arguments{
\item{object}{A \code{estim.bin} object}
}
\value{
model string
}
\description{
Get Model Name
}
