% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\docType{class}
\name{evaluation-class}
\alias{evaluation}
\alias{evaluation-class}
\title{Represents a scoring rule}
\section{Fields}{

\describe{
\item{\code{Name}}{Abbreviation of the rule}

\item{\code{Description}}{Full name and any other description about the rule}

\item{\code{IsPositivelyOriented}}{A boolean indicates wether higher is better}

\item{\code{Score}}{The scoring function. It should get two arguments: vector of forecast errors and vector of forecast standard errors and it should return a vector of the same length}
}}

