% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{exogenous}
\alias{exogenous}
\title{Extract Exogenous Variable(s) Data}
\usage{
exogenous(object, equation = 1, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}.}

\item{equation}{A number or a string specifying the equation with exogenous data.}

\item{...}{Additional arguments.}
}
\value{
A matrix containing the exogenous data.
}
\description{
This function extracts data of an exogenous variable(s) in an equation from an estimated model.
It takes zero restrictions imposed into account.
}
