% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_CrossSection}
\alias{F_CrossSection}
\title{Creates a Cross-Section Frequency}
\usage{
F_CrossSection(position)
}
\arguments{
\item{position}{Position of the observation}
}
\value{
An object of class 'ldtf'
}
\description{
This frequency is generally for indexed (or, non-time-series) data. It is an integer that represents the position of the observation.
}
\details{
\itemize{
\item \strong{Value String:} \code{"#"} (number is \code{position})
\item \strong{Class String:} \code{"cs"}
}
}
