% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Daily}
\alias{F_Daily}
\title{Creates a \code{Daily} Frequency}
\usage{
F_Daily(year, month, day)
}
\arguments{
\item{year}{Year of the observation}

\item{month}{Month of the observation}

\item{day}{Day of the observation.}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every day
}
\details{
\itemize{
\item \strong{Value String:} \code{"YYYYMMDD"} (similar to \code{Weekly})
\item \strong{Class String:} \code{"d"}
}
}
