% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Hourly}
\alias{F_Hourly}
\title{Creates an 'Hourly' Frequency}
\usage{
F_Hourly(day, hour)
}
\arguments{
\item{day}{A 'Day-based' frequency such as \code{Daily} or \code{Daily-In-Week}}

\item{hour}{Index of hour in the day (1 to 24)}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every hour
}
\details{
\itemize{
\item \strong{Value String:} \code{"YYYYMMDD:#"} (the number is \code{hour})
\item \strong{Class String:} \code{ho|...} (the ... is the 'Class String' of \code{day})
}
}
