% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_ListString}
\alias{F_ListString}
\title{Creates an \code{List-String} Frequency}
\usage{
F_ListString(items, value)
}
\arguments{
\item{items}{Items of the list}

\item{value}{Current item}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that is labeled by string
}
\details{
\itemize{
\item \strong{Value String:} \code{"..."} (in which ... is the \code{value})
\item \strong{Class String:} \code{Ls} or \code{Ls:...} (in which ...
is the semi-colon separated \code{items})
}
}
