% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Minute_ly}
\alias{F_Minute_ly}
\title{Creates an 'Minute-ly' Frequency}
\usage{
F_Minute_ly(day, minute)
}
\arguments{
\item{day}{A 'Day-based' frequency such as daily or daily-in-week}

\item{minute}{Index of Minute in the day (1 to 1440)}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every minute
}
\details{
\itemize{
\item \strong{Value String:} \code{"YYYYMMDD:#"} (the number is \code{minute})
\item \strong{Class String:} \code{mi|...} (the ... is the 'Class String' of \code{day})
}
}
