% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Monthly}
\alias{F_Monthly}
\title{Creates a \code{Monthly} Frequency}
\usage{
F_Monthly(year, month)
}
\arguments{
\item{year}{Year of the observation}

\item{month}{Month of the observation}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every month
}
\details{
\itemize{
\item \strong{Value String:} \code{"#m#"} (first # is the \code{year}, second # is
\code{month} (1 to 12); e.g., 2010m8 or 2010m12.
Note that 2000m0 or 2000m13 are invalid.
\item \strong{Class String:} \code{"m"}
}
}
