% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_MultiDaily}
\alias{F_MultiDaily}
\title{Creates an \code{Multi-Daily} Frequency}
\usage{
F_MultiDaily(year, month, day, k)
}
\arguments{
\item{year}{Year of the observation}

\item{month}{Month of the observation}

\item{day}{First day of the observation}

\item{k}{Number of the days}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every \code{k} days
}
\details{
\itemize{
\item \strong{Value String:} \code{"YYYYMMDD"} (similar to \code{Weekly})
\item \strong{Class String:} \code{"d#"} (the number is \code{k})
}
}
