% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_MultiWeekly}
\alias{F_MultiWeekly}
\title{Creates a \code{Multi-Weekly} Frequency}
\usage{
F_MultiWeekly(year, month, day, k)
}
\arguments{
\item{year}{Year of the observation}

\item{month}{Month of the observation}

\item{day}{First day of the observation. It points to the first day of the week}

\item{k}{Number of weeks}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every 'k' weeks
}
\details{
\itemize{
\item \strong{Value String:} \code{"YYYYMMDD"} (similar to \code{Weekly})
\item \strong{Class String:} \code{"w#"} (the number is \code{k}; e.g., w3 means every 3 weeks)
}
}
