% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_MultiYearly}
\alias{F_MultiYearly}
\title{Creates a \code{Multi-Yearly} Frequency}
\usage{
F_MultiYearly(year, z)
}
\arguments{
\item{year}{Year of the observation}

\item{z}{Number of years}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every \code{z} years
}
\details{
\itemize{
\item \strong{Value String:} \code{"#"} (similar to \code{Yearly})
\item \strong{Class String:} \code{"z#"} (integer represents the
\code{z}; e.g., z3)
}
}
