% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Quarterly}
\alias{F_Quarterly}
\title{Creates a \code{Quarterly} Frequency}
\usage{
F_Quarterly(year, quarter)
}
\arguments{
\item{year}{Year of the observation}

\item{quarter}{Quarter of the observation (1 to 4)}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every quarter
}
\details{
\itemize{
\item \strong{Value String:} \code{"#q#"} (first # is \code{year}, second # is
\code{quarter}; e.g., 2010q3 or 2010q4.
Note that 2000q0 or 2000q5 are invalid.
\item \strong{Class String:} \code{"q"}
}
}
