% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Second_ly}
\alias{F_Second_ly}
\title{Creates an 'Second-ly' Frequency}
\usage{
F_Second_ly(day, second)
}
\arguments{
\item{day}{A 'Day-based' frequency such as daily or daily-in-week}

\item{second}{Index of second in the day (1 to 86400)}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every second
}
\details{
\itemize{
\item \strong{Value String:} \code{"YYYYMMDD:#"} (the number is \code{second})
\item \strong{Class String:} \code{se|...} (the ... is the 'Class String' of \code{day})
}
}
