% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Weekly}
\alias{F_Weekly}
\title{Creates a \code{Weekly} Frequency}
\usage{
F_Weekly(year, month, day)
}
\arguments{
\item{year}{Year of the observation}

\item{month}{Month of the observation}

\item{day}{Day of the observation. It points to the first day of the week}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every week
}
\details{
\itemize{
\item \strong{Value String:} \code{"YYYYMMDD"} (\code{YYYY} is the \code{year}, \code{MM} is \code{month} and \code{DD} is \code{day})
\item \strong{Class String:} \code{"w"}
}
}
