% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_XTimesADay}
\alias{F_XTimesADay}
\title{Creates an 'X-Times-A-Day' Frequency}
\usage{
F_XTimesADay(day, x, position)
}
\arguments{
\item{day}{A 'Day-based' frequency such as daily or daily-in-week}

\item{x}{Number of observations in a day}

\item{position}{Current position}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens x times in a day
}
\details{
\itemize{
\item \strong{Value String:} \code{"#"} (the number is \code{hour})
\item \strong{Class String:} \code{"da#|..."} (the number is \code{x}
and ... is the 'Class String' of \code{day}))
}
}
