% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_XTimesAYear}
\alias{F_XTimesAYear}
\title{Creates an \code{X-Times-A-Year} Frequency}
\usage{
F_XTimesAYear(year, x, position)
}
\arguments{
\item{year}{Year of the observation}

\item{x}{Number of observation in each year}

\item{position}{Position of the current observation}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens \code{x} times every year
}
\details{
\itemize{
\item \strong{Value String:} \code{"#:#"} (first # is \code{year} and second # is
\code{position};
e.g., 2010:8/12 or 2010:10/10.
Note that 2000:0/2 or 2000:13/12 are invalid.
\item \strong{Class String:} \code{"y#"} (the number is \code{x})
}
}
