% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_XTimesZYear}
\alias{F_XTimesZYear}
\title{Creates an \code{X-Times-Z-Years} Frequency}
\usage{
F_XTimesZYear(year, x, z, position)
}
\arguments{
\item{year}{Year of the observation}

\item{x}{Number of partitons in each z years}

\item{z}{Number of years}

\item{position}{Position of the current observation}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens \code{x} times each \code{z} years
}
\details{
\itemize{
\item \strong{Value String:} \code{"#:#"} (Similar to \code{X-Times-A-Year})
\item \strong{Class String:} \code{"x#z#"} (first # is \code{x}, second # is \code{z};
e.g., x23z4 means 23 times every 4 years)
}
}
