% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_Yearly}
\alias{F_Yearly}
\title{Creates a \code{Yearly} Frequency}
\usage{
F_Yearly(year)
}
\arguments{
\item{year}{Year of the observation}
}
\value{
An object of class 'ldtf'
}
\description{
Frequency for a series that happens every year
}
\details{
\itemize{
\item \strong{Value String:} \code{"#"} (number is \code{year})
\item \strong{Class String:} \code{"y"}
}
}
