% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{fitted.ldt.estim}
\alias{fitted.ldt.estim}
\title{Extract Fitted Data}
\usage{
\method{fitted}{ldt.estim}(object, equations = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}.}

\item{equations}{A number, a numeric array or a string array specifying the equations with residual data. If \code{NULL}, residuals in all equations are returned.}

\item{...}{Additional arguments.}
}
\value{
A matrix containing the exogenous data.
}
\description{
This function calculates and returns fitted values for an \code{ldt.estim} object.
}
