% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get.data.append.newX}
\alias{get.data.append.newX}
\title{Append \code{newX} to \code{data$data} matrix.}
\usage{
get.data.append.newX(data, maxHorizon = NA)
}
\arguments{
\item{data}{The output of \link{get.data} function.}

\item{maxHorizon}{Number of expected new data}
}
\value{
The input \code{data} with updated data matrix
}
\description{
Use it for VARMA estimation
}
