% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get.data.check.discrete}
\alias{get.data.check.discrete}
\title{Check if a column is discrete}
\usage{
get.data.check.discrete(data, colIndex = 1)
}
\arguments{
\item{data}{Output of \link{get.data} function}

\item{colIndex}{The index of column to be checked.}
}
\value{
Number of choices in the model, if no error occured.
This means, the maximum value for the discrete data will be the output minus one.
}
\description{
For example, it checks if the endogenous variable in binary model is 0 and 1 (number of choices is 2)
}
