% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get.data.check.intercept}
\alias{get.data.check.intercept}
\title{Check for an intercept in a matrix}
\usage{
get.data.check.intercept(matrix)
}
\arguments{
\item{matrix}{data matrix}
}
\value{
The index of the intercept. '-1' in intercept is not found.
}
\description{
This function checks if any column in the matrix is intercept.
}
