% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get.data.keep.complete}
\alias{get.data.keep.complete}
\title{Remove Rows with Missing Observations from Data}
\usage{
get.data.keep.complete(data, warn = TRUE)
}
\arguments{
\item{data}{Output of \link{get.data} function}

\item{warn}{If true,  warning message about the indices of the removed rows is shown}
}
\value{
The input \code{data} but with updated \code{data$data} and \code{data$obsCount}
}
\description{
Remove Rows with Missing Observations from Data
}
