% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get.search.options}
\alias{get.search.options}
\title{Get Extra Options for Model Search Process}
\usage{
get.search.options(parallel = FALSE, reportInterval = 0)
}
\arguments{
\item{parallel}{If \code{TRUE}, a parallel search algorithm is used. This generally changes the speed and memory usage.}

\item{reportInterval}{An integer representing the time interval (in seconds) for reporting progress (if any significant change has occurred). Set to zero to disable reporting.}
}
\value{
A list with the given options.
}
\description{
Use this function to determine how the model search is performed.
}
