% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetAuc}
\alias{GetAuc}
\title{Gets the Area Under the receiver Operating Characteristic (ROC) Curve}
\usage{
GetAuc(y, scores, weights = NULL)
}
\arguments{
\item{y}{(numeric vector) actual values.}

\item{scores}{(numeric matrix) a matrix with scores in the columns.}

\item{weights}{(numeric vector) weights of the observations}
}
\value{
value of the AUC
}
\description{
Gets the Area Under the receiver Operating Characteristic (ROC) Curve
}
\examples{
y = c(0,1,0,1,0,1,0,1)
scores = matrix(c( 0.4, 0.6, 0.45, 0.9, 0.6, 0.3, 0.5, 0.1,
                   0.6, 0.4, 0.55, 0.1, 0.4, 0.7, 0.5, 0.9), 8, 2)
res = GetAuc(y,scores,NULL)
}
