% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetCombination4Moments}
\alias{GetCombination4Moments}
\title{Combines Two Distributions Defined by their First 4 Moments}
\usage{
GetCombination4Moments(mix1, mix2)
}
\arguments{
\item{mix1}{(list) First distribution which is defined by a list with mean, variance, skewness, kurtosis, sumWeights, count}

\item{mix2}{(list) Second distribution (similar to \code{mix1}).}
}
\value{
(list) A list similar to \code{mix1}
}
\description{
Combines Two Distributions Defined by their First 4 Moments
}
\examples{
#see its \code{test_that} function
}
