% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDistance}
\alias{GetDistance}
\title{Gets Distances Between Variables}
\usage{
GetDistance(
  data,
  distance = "correlation",
  correlation = "pearson",
  checkNan = TRUE
)
}
\arguments{
\item{data}{(numeric matrix) Data with variables in the columns.}

\item{distance}{(string) Determines how distances are calculated. It can be \code{correlation}, \code{absCorrelation}, \code{euclidean}, \code{manhattan}, \code{maximum}.}

\item{correlation}{(string) If \code{distance} is correlation, it determines the type of the correlation. It can be \code{pearson}, \code{spearman}.}

\item{checkNan}{(bool) If false, \code{NAN}s are not omitted.}
}
\value{
A symmetric matrix (lower triangle as a vector).
}
\description{
Gets Distances Between Variables
}
