% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{GetEstim}
\alias{GetEstim}
\title{Get Estimation from Search Result}
\usage{
GetEstim(searchRes, endoIndices, exoIndices, y, x, printMsg, ...)
}
\arguments{
\item{searchRes}{an object of class \code{ldtsearch}}

\item{endoIndices}{endogenous indices}

\item{exoIndices}{exogenous indices}

\item{y}{dependent variables data}

\item{x}{exogenous variables data}

\item{printMsg}{argument to be passed to the estimation methods}

\item{...}{additional arguments}
}
\value{
estimation result
}
\description{
Get Estimation from Search Result
}
