% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetMeasureFromWeight}
\alias{GetMeasureFromWeight}
\title{Converts a Measure to Weight}
\usage{
GetMeasureFromWeight(value, measureName)
}
\arguments{
\item{value}{(double) the measure}

\item{measureName}{(string) measure name}
}
\value{
the measure
}
\description{
Converts a Measure to Weight
}
\examples{
weight <- GetWeightFromMeasure(-3.4, "sic")
measure <- GetMeasureFromWeight(weight, "sic")
}
