% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetNelderMeadOptions}
\alias{GetNelderMeadOptions}
\title{Options for Nelder-Mead Optimization}
\usage{
GetNelderMeadOptions(
  maxIterations = 100L,
  epsilon = 1e-08,
  alpha = 1,
  beta = 0.5,
  gamma = 2,
  scale = 1
)
}
\arguments{
\item{maxIterations}{(int) Maximum number of iterations.}

\item{epsilon}{(double) A small value to test convergence.}

\item{alpha}{(double) the reflection coefficient.}

\item{beta}{(double) the contraction coefficient.}

\item{gamma}{(double) the expansion coefficient.}

\item{scale}{(double) A scale in initializing the simplex.}
}
\value{
A list with the given options.
}
\description{
Options for Nelder-Mead Optimization
}
