% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetNewtonOptions}
\alias{GetNewtonOptions}
\title{Options for Newton Optimization}
\usage{
GetNewtonOptions(
  maxIterations = 100L,
  functionTol = 1e-04,
  gradientTol = 0,
  useLineSearch = TRUE
)
}
\arguments{
\item{maxIterations}{(int) Maximum number of iterations.}

\item{functionTol}{(double) A small value to test convergence of the objective function.}

\item{gradientTol}{(double) A small value to test convergence of the gradient.}

\item{useLineSearch}{(bool) If true, it uses line search.}
}
\value{
A list with the given options.
}
\description{
Options for Newton Optimization
}
