% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPcaOptions}
\alias{GetPcaOptions}
\title{Options for PCA}
\usage{
GetPcaOptions(ignoreFirst = 1L, exactCount = 0L, cutoffRate = 0.8, max = 1000L)
}
\arguments{
\item{ignoreFirst}{(int) Excludes variables at the beginning of data matrices (such as intercept) from PCA.}

\item{exactCount}{(int) Determines the number of components to be used. If zero, number of components are determined by the \code{cutoffRate}.}

\item{cutoffRate}{(double between 0 and 1) Determines the cutoff rate for cumulative variance ratio in order to determine the number of PCA components. It is not used if \code{exactCount} is positive.}

\item{max}{(int) Maximum number of components when \code{cutoffRate} is used.}
}
\value{
A list with the given options.
}
\description{
Options for PCA
}
