% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetWeightFromMeasure}
\alias{GetWeightFromMeasure}
\title{Converts a Measure to Weight}
\usage{
GetWeightFromMeasure(value, measureName)
}
\arguments{
\item{value}{(double) the measure}

\item{measureName}{(string) measure name}
}
\value{
the weight
}
\description{
Converts a Measure to Weight
}
\examples{
weight <- GetWeightFromMeasure(-3.4, "sic")
}
