% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GldDensityQuantile}
\alias{GldDensityQuantile}
\title{Gets GLD Density Quantile}
\usage{
GldDensityQuantile(data, L1, L2, L3, L4)
}
\arguments{
\item{data}{(numeric vector) data}

\item{L1}{(double) First parameter}

\item{L2}{(double) Second parameter}

\item{L3}{(double) Third parameter}

\item{L4}{(double) Fourth parameter}
}
\value{
(numeric vector) result
}
\description{
Gets GLD Density Quantile
}
