% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{h.get.estim}
\alias{h.get.estim}
\title{Gets Estimation from Search Result}
\usage{
h.get.estim(searchRes, endoIndices, exoIndices, y, x, printMsg, ...)
}
\arguments{
\item{searchRes}{Output from one of the \code{search.?} functions (see \link{search.sur}, \link{search.varma}, or \link{search.bin}).}

\item{endoIndices}{Endogenous indices corresponding to the columns of the dependent variable.}

\item{exoIndices}{Exogenous indices corresponding to the columns of the exogenous variable.}

\item{y}{Data for dependent variables.}

\item{x}{Data for exogenous variables.}

\item{printMsg}{Argument passed to estimation methods}

\item{...}{Additional arguments}
}
\value{
Estimation result similar to output of \link{estim.sur}, \link{estim.varma}, or \link{estim.bin}.
}
\description{
This function retrieves the estimation result from the output of a search process.
It uses the search options to re-estimate the same model estimated in the search process,
given a list of required indices.
}
\seealso{
\link{estim.sur}, \link{estim.varma}, \link{estim.bin}
}
