% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{IsEmailValid}
\alias{IsEmailValid}
\title{Determines if an email address is valid (this is not exact. Just use it to avoid mistakes)}
\usage{
IsEmailValid(x)
}
\arguments{
\item{x}{email}
}
\value{
\code{TRUE} if email is valid, \code{FALSE} otherwise.
}
\description{
Determines if an email address is valid (this is not exact. Just use it to avoid mistakes)
}
