% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{IsGuidValid}
\alias{IsGuidValid}
\title{Determines if a GUID is valid}
\usage{
IsGuidValid(x)
}
\arguments{
\item{x}{GUID}
}
\value{
\code{TRUE} if GUID is valid, \code{FALSE} otherwise.
}
\description{
Determines if a GUID is valid
}
