% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldt.R
\docType{class}
\name{ldt-class}
\alias{ldt}
\alias{ldt-class}
\title{The main class in ldt package.}
\details{
This is a reference class. Use $new to generate an instance.
example:
load("data/endodata_rand.rda")
load("data/exodata_rand.rda")
a=ldt$new(endodata,exodata, maxsize = 3)
a$Run()
}
\section{Fields}{

\describe{
\item{\code{EndoData}}{A ts that contains the data for the endogenous variable. The target variable (the one we are trying to forecast) is the first variable in this series.}

\item{\code{ExoData}}{A matrix that contains the data for the exogenous variable. note that the frequency and start date will be fetched from EndoData}

\item{\code{MaxHorizon}}{A positive integer that indicates the maximum of the forecast horizons. Forecasts will be provided up to this horizon. Default is 1}

\item{\code{MaxLag}}{A positive integer that indicates the maximum lag of models (including AR and MA parts). Default is 1}

\item{\code{MaxSize}}{A positive integer that indicates the maximum size of multiple/multivariate models. Default is 1. For example, in arima case, a value of 2 means that models might have zero or one exogenous variables.}

\item{\code{ScoringRules}}{A list of evaluation objects. Default is a list of AIC, BIC, MAE, MSE, LSR, LnSR, QSR, HSR and CRPSR (for all score rules, normality is assumed)}

\item{\code{SimulationCount}}{A positive integer that indicates the number of Out-of-sample evaluations. i.e., how many times we should seperate the data into training and evaluation samples and test the forecast accuracy of different models. The default is 1}

\item{\code{Packs}}{A list of ldtpack objects}
}}

