% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldt.R
\name{ldt_initialize}
\alias{ldt_initialize}
\title{The constructor of ldt class}
\arguments{
\item{endodata}{sets EndoData field of the class}

\item{exodata}{sets ExoData field of the class}

\item{maxhorizon}{sets MaxHorizon field of the class}

\item{maxlag}{sets MaxLag field of the class}

\item{maxsize}{sets MaxSize field of the class}

\item{simulationcount}{sets SimulationCount field of the class}
}

