% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldtpack.R
\docType{class}
\name{ldtpack-class}
\alias{ldtpack}
\alias{ldtpack-class}
\title{Represents a specific forecast package}
\section{Fields}{

\describe{
\item{\code{ParentLDT}}{The main ldt class}

\item{\code{ID}}{A unique string that represents the related package}

\item{\code{Description}}{A description for the related package}

\item{\code{SupportedScoringRules}}{A list of 'evaluation' objects that this package can provide. Not all packages can report (e.g.) AIC}

\item{\code{Processes}}{A list of ldtpacksub}

\item{\code{Results}}{A list that contains the results for each member of SupportedScoringRules.}

\item{\code{AllModels}}{Just in case you wanted to check all the models}
}}

