% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldtpack.R
\name{ldtpack_considernew}
\alias{ldtpack_considernew}
\title{consider a new score}
\arguments{
\item{model}{the model based on which the scores are generated}

\item{i}{the index of evalation in SupportedScoringRules based on which the scores are generated}

\item{scores}{the generated scores}
}

