% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{logLik.ldt.estim}
\alias{logLik.ldt.estim}
\title{Extract Maximum Log-Likelihood}
\usage{
\method{logLik}{ldt.estim}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}}

\item{...}{Additional arguments.}
}
\value{
The value of the maximum log-likelihood for the whole system.
}
\description{
This function extracts maximum log-likelihood from an \code{ldt.estim} object.
}
