% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Parse_F}
\alias{Parse_F}
\title{Converts back a String to \code{ldtf} Object}
\usage{
Parse_F(str, classStr)
}
\arguments{
\item{str}{value of the frequency. It must be an \code{ldtf}
object returned from \code{F_?} functions.}

\item{classStr}{class of the frequency}
}
\value{
An object of class 'ldtf'
}
\description{
The format is explained in \code{F_?} functions.
}
