% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.plots.R
\name{plot.ldt.varma.prediction}
\alias{plot.ldt.varma.prediction}
\title{Plot Predictions from a VARMA model}
\usage{
\method{plot}{ldt.varma.prediction}(
  x,
  variable = 1,
  xAxisArgs = list(),
  fanPlotArgs = list(),
  simMetric = NULL,
  simLineArgs = list(),
  simPointsArgs = list(),
  ...
)
}
\arguments{
\item{x}{An object of class \code{ldt.varma.prediction}, which is the output of \link{predict.ldt.estim.varma} function.}

\item{variable}{Index or name of the variable to be plotted.}

\item{xAxisArgs}{Arguments to pass to \code{axis} function}

\item{fanPlotArgs}{Additional arguments for the \link{fan.plot} function. \code{lambda} is added automatically.}

\item{simMetric}{Name of metric to plot its details, provided that simulation details are available. If \code{NULL}, simulation details are not plotted.}

\item{simLineArgs}{Arguments to pass to \code{line} function for simulation lines (if available).}

\item{simPointsArgs}{Arguments to pass to \code{points} function for simulation points (if available).}

\item{...}{Additional parameters (unused)}
}
\value{
This function does not return a value.
}
\description{
Plot Predictions from a VARMA model
}
