% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{predict.ldt.estim}
\alias{predict.ldt.estim}
\title{Extract Prediction Results}
\usage{
\method{predict}{ldt.estim}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim}}

\item{...}{Additional arguments.}
}
\value{
A list containing the predicted (projected) means and variances.
}
\description{
This function extracts predicted mean and its variance from an \code{ldt.estim} object.
new data must be provided while estimating the model.
}
