% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.regression.R
\name{predict.ldt.estim.varma}
\alias{predict.ldt.estim.varma}
\title{Extract Prediction Results from a \code{ldt.estim.varma} Object}
\usage{
\method{predict}{ldt.estim.varma}(object, actualCount = 0, startFrequency = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ldt.estim.varma}}

\item{actualCount}{Number of actual observations to be included in the result.}

\item{startFrequency}{Frequency of the first observation used in the estimation.
This is object of class \code{ldtf}.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{ldt.varma.prediction}, which is a list with predicted \code{means} and (if available) \code{variances}.
}
\description{
This function extracts predicted mean and its variance from an \code{ldt.estim.varma} object.
new data must be provided while estimating the model.
}
