% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.prints.R
\name{print.ldt.estim.projection}
\alias{print.ldt.estim.projection}
\title{Prints an \code{ldt.estim.projection} object}
\usage{
\method{print}{ldt.estim.projection}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ldt.estim.projection}}

\item{...}{Additional arguments}
}
\value{
This function has no output.
}
\description{
An \code{ldt.estim.projection} object is the output of \code{\link[=predict.ldt.estim]{predict.ldt.estim()}} function.
}
