% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.prints.R
\name{print.ldt.estim}
\alias{print.ldt.estim}
\title{Prints an \code{ldt.estim} object}
\usage{
\method{print}{ldt.estim}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ldt.estim}}

\item{...}{Additional arguments}
}
\value{
This function has no output.
}
\description{
Prints the main results in an \code{ldt.estim} object.
}
\details{
An \code{ldt.search} object is an output from one of the \code{search.?} functions (see \code{search.sur}, \code{search.varma}, or \code{search.bin}).
}
