% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.prints.R
\name{print.ldt.list}
\alias{print.ldt.list}
\title{Prints an \code{ldt.list} object}
\usage{
\method{print}{ldt.list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ldt.list}}

\item{...}{Additional arguments}
}
\value{
This function has no output.
}
\description{
Prints an \code{ldt.list} object
}
