% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.prints.R
\name{print.ldt.varma.prediction}
\alias{print.ldt.varma.prediction}
\title{Prints an \code{ldt.varma.prediction} object}
\usage{
\method{print}{ldt.varma.prediction}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ldt.varma.prediction}}

\item{...}{Additional arguments}
}
\value{
This function has no output.
}
\description{
An \code{ldt.varma.prediction} object is the output of \code{\link[=predict.ldt.estim.varma]{predict.ldt.estim.varma()}} function.
}
