% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.ldtsearch}
\alias{print.ldtsearch}
\title{Prints the Output of a Search Process}
\usage{
\method{print}{ldtsearch}(x, ...)
}
\arguments{
\item{x}{An output from one of the \code{search.?} functions (see \code{search.sur}, \code{search.varma}, or \code{search.bin}).}

\item{...}{Additional arguments}
}
\value{
This function has no output.
}
\description{
Prints the Output of a Search Process
}
