% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.ldtv}
\alias{print.ldtv}
\title{Prints an \code{ldtv} object}
\usage{
\method{print}{ldtv}(x, ...)
}
\arguments{
\item{x}{An \code{ldtv} object}

\item{...}{additional arguments}
}
\value{
\code{NULL}
}
\description{
Prints an \code{ldtv} object
}
