% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{search.steps}
\alias{search.steps}
\title{Step-wise estimation}
\usage{
search.steps(method, isInnerExogenous, ...)
}
\arguments{
\item{method}{sur, bin or varma}

\item{isInnerExogenous}{Determines if the inner indices are for exogenous variables.}

\item{...}{Additional arguments for the search function.}
}
\value{
the result
}
\description{
This function uses the calculated inclusion weights and selects a subset of variables in each step.
Note that it uses the values for the first target variable and first metric and might not be suitable for multi-target or multi-metric searches.
}
