% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sequence_F}
\alias{Sequence_F}
\title{Generates a Sequence for a frequency}
\usage{
Sequence_F(start, length)
}
\arguments{
\item{start}{first element of the sequence. It must be an \code{ldtf}
object returned from \code{F_?} functions.}

\item{length}{Length of the sequence}
}
\value{
A list of strings
}
\description{
Generates a Sequence for a frequency
}
