% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.summary.R
\name{summary.ldt.search.item}
\alias{summary.ldt.search.item}
\title{Summary for an \code{ldt.search.item} object}
\usage{
\method{summary}{ldt.search.item}(object, searchResult = NULL, test = FALSE, ...)
}
\arguments{
\item{object}{An \code{ldt.search.item} object.}

\item{searchResult}{Parent list of \code{object}, which is an \code{ldt.search} object.}

\item{test}{If \code{TRUE} and applicable (e.g., in model estimation), it checks the metrics and throws error for any inconsistencies between the current estimation and the one calculated in the search process.}

\item{...}{Additional arguments.}
}
\value{
If the object contains the indices of endogenous variables of an estimated model, it returns the estimation output.
Otherwise, it returns object.
}
\description{
While you can get a summary of an item in a search result, this function is mainly designed to be called from \link{print.ldt.search} function.
Its main job is to estimate the full model using the reported indices from the search process.
}
\details{
An \code{ldt.search.item} object is a member of \code{ldt.search} object.
An \code{ldt.search} object is an output from one of the \code{search.?} functions (see \code{search.sur}, \code{search.varma}, or \code{search.bin}).
}
