% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.summary.R
\name{summary.ldt.search}
\alias{summary.ldt.search}
\title{Summary for an \code{ldt.search} object}
\usage{
\method{summary}{ldt.search}(object, test = FALSE, ...)
}
\arguments{
\item{object}{An \code{ldt.search} object.}

\item{test}{If \code{TRUE} and applicable (e.g., in model estimation), it checks the metrics and throws error for any inconsistencies between the current estimation and the one calculated in the search process (Provided that negative seed is used).}

\item{...}{Additional arguments.}
}
\value{
The output replaces the value of \code{object$results} with the summary from \link{summary.ldt.search.item}.
}
\description{
Use this function to get the full estimation of the models reported in the output of a search process.
}
\details{
An \code{ldt.search} object is an output from one of the \code{search.?} functions (see \code{search.sur}, \code{search.varma}, or \code{search.bin}).
}
