% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ToString_F}
\alias{ToString_F}
\title{Converts an \code{ldtf} Object to String}
\usage{
ToString_F(value)
}
\arguments{
\item{value}{value of the frequency. It must be an \code{ldtf}
object returned from \code{F_?} functions.}
}
\value{
An object of class 'ldtf'
}
\description{
The format is explained in \code{F_?} functions.
}
