% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vig_data}
\alias{vig_data}
\title{Data for Vignettes (and Tests)}
\format{
A list
}
\description{
A subset of different data sets generally for tests and vignettes.
Data is generated from \code{Data_?} functions.
}
\details{
\itemize{
\item wdi. data from WDI data set.
\item berka. data from Berka data set.
\item vesta. data from Vesta data set.
\item pcp. data from PCP data set.
}
}
