% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.ldtsearch}
\alias{print.ldtsearch}
\title{Print an \code{ldtsearch} object}
\usage{
\method{print}{ldtsearch}(x, ...)
}
\arguments{
\item{x}{\code{ldtsearch} object}

\item{...}{additional arguments}
}
\value{
\code{NULL}
}
\description{
Print an \code{ldtsearch} object
}
