% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Search_s}
\alias{Search_s}
\title{Stepwise estimation}
\usage{
Search_s(
  method,
  data,
  sizes = list(c(1, 2), c(3, 4), c(5), c(6:10)),
  counts = c(NA, 40, 30, 20),
  savePre,
  printMsg = FALSE,
  ...
)
}
\arguments{
\item{method}{sur, dc or varma}

\item{data}{exogenous (for sur and dc) or endogenous (for varma)}

\item{sizes}{detemines the steps}

\item{counts}{determines the size in each step}

\item{savePre}{if not \code{NULL}, it saves and tries to load the
progress of search step in a file (name=\code{paste0(savePre,i)} where
\code{i} is the index of the step).}

\item{printMsg}{If true, some information about the steps is printed.
Note that it is different from searchers' \code{printMsg}.}

\item{...}{Additional arguments}
}
\value{
the result
}
\description{
Stepwise estimation
}
